bookstore/requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
pymongo==4.6.2
